--- packages/core/src/lib/default-callbacks.ts
@@ -1,18 +1,0 @@
-import type { CallbacksOptions } from "../types.js"
-
-export const defaultCallbacks: CallbacksOptions = {
-  signIn() {
-    return true
-  },
-  redirect({ url, baseUrl }) {
-    if (url.startsWith("/")) return `${baseUrl}${url}`
-    else if (new URL(url).origin === baseUrl) return url
-    return baseUrl
-  },
-  session({ session }) {
-    return session
-  },
-  jwt({ token }) {
-    return token
-  },
-}
