+++ packages/core/src/lib/oauth/authorization-url.ts
@@ -6,7 +6,7 @@ import type {
   RequestInternal,
   ResponseInternal,
 } from "../../types.js"
-import type { Cookie } from "../cookie.js"
+import type { Cookie } from "../utils/cookie.js"
 
 /**
  * Generates an authorization/request token URL.
