+++ packages/core/src/lib/routes/callback/handle-login.ts
@@ -1,15 +1,15 @@
-import { OAuthAccountNotLinked } from "../errors.js"
-import { fromDate } from "./utils/date.js"
+import { OAuthAccountNotLinked } from "../../../errors.js"
+import { fromDate } from "../../utils/date.js"
 
 import type {
   AdapterAccount,
   AdapterSession,
   AdapterUser,
-} from "../adapters.js"
-import type { Account, InternalOptions, User } from "../types.js"
-import type { JWT } from "../jwt.js"
-import type { OAuthConfig } from "../providers/index.js"
-import type { SessionToken } from "./cookie.js"
+} from "../../../adapters.js"
+import type { Account, InternalOptions, User } from "../../../types.js"
+import type { JWT } from "../../../jwt.js"
+import type { OAuthConfig } from "../../../providers/index.js"
+import type { SessionToken } from "../../utils/cookie.js"
 
 /**
  * This function handles the complex flow of signing users in, and either creating,
