+++ packages/core/src/lib/routes/callback/oauth/csrf-token.ts
@@ -1,6 +1,6 @@
-import { createHash, randomString } from "./utils/web.js"
+import { createHash, randomString } from "../../../utils/web.js"
 
-import type { InternalOptions } from "../types.js"
+import type { InternalOptions } from "../../../../types.js"
 interface CreateCSRFTokenParams {
   options: InternalOptions
   cookieValue?: string
