+++ packages/core/src/lib/routes/index.ts
@@ -1,4 +1,4 @@
-export { callback } from "./callback.js"
+export { callback } from "./callback"
 export { providers } from "./providers.js"
 export { session } from "./session.js"
 export { signin } from "./signin.js"
